+++ Meraki.Api/Data/SmNetworkUser.cs
@@ -6,7 +6,7 @@
 [ApiAccessReadOnlyClass]
 [DataContract]
 
-public class SmNetworkUsers
+public class SmNetworkUser
 {
 	/// <summary>
 	/// Id
@@ -18,13 +18,13 @@ public class SmNetworkUsers
 	/// Email
 	/// </summary>
 	[DataMember(Name = "email")]
-	public string Email { get; set; } = string.Empty;
+	public string? Email { get; set; }
 
 	/// <summary>
 	/// Full name
 	/// </summary>
 	[DataMember(Name = "fullName")]
-	public string FullName { get; set; } = string.Empty;
+	public string? FullName { get; set; }
 
 	/// <summary>
 	/// Username
@@ -90,5 +90,5 @@ public class SmNetworkUsers
 	/// User thumbnail
 	/// </summary>
 	[DataMember(Name = "userThumbnail")]
-	public string UserThumbnail { get; set; } = string.Empty;
+	public string? UserThumbnail { get; set; }
 }