+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Changed AdaptivePolicyPolicies to correct nullability. Changed timestamps to DateTime
 - Updated RoutingOspf to add RoutingOspfV3 settings
 - Made all DeviceClient properties null
+- Updated nullability on SmNetworkUser
 
 ## v1.20.42
 - v1.15 changes